Recommend unique filename for new/default score
In the special case of a newly created score that has a default
title, and there already exists an "Untitled\ score.mscz" file
in the chosen save directory, recommend a different file name
when saving (ex. "Untitled\ score (1).mscz") to prevent
accidental overwriting.

Inspired by when you download a file with the same name from the
web, it saves a separate copy instead of overwriting the existing
file.

Issue: #20140